Create README.md
For our assignment in order to apply what we learned in class, we were asked to create the CSS box model.  The box model is referenced when creating a layout.  According to W3 Schools, "it's a box that wraps around every HTML element, and contains margins, borders, padding, and content."  Creating this assignment helped me to learn and understand the use of spacing, padding, width, alignment, overflow, breaking up content into sections.  It was an easy challenge for me to do from scratch and the first complete draft took me 40 minutes to complete.  Afterwards I spent another two sessions to finalize the assignment and fine tune any small details or issues with overflow.